FileComparator now works bit-wise.
Sort mode is constructed using or operations on several options - sort by (name or last modified), group folders (yes or no) and sort order ("normal" (A-Z) or reversed). The menu will have to be connected to the preferences, as there is currently no way to change these values in settings and there are no default values shown/ticked.