Toolbar menu options and settings fragment are now connected.
Preferences are saved as strings and static methods from the settings fragment are used to "translate" them. There is a special static variable inside the settings fragment class which is used to tell when the preferences were changed. This value is checked in onResume to update the LiveData and sync the menu.