Show progress bar when loading large lists of files.
There are a few issues though - the progress bar does not have any in/out animation (TransitionManager does not work good, as the two in/out transitions are too closely together and the view remains visible). Furthermore, after showing the recycler view again, the old items can be seen for a split second as the DIFF gets calculated. The progress bar is also not shown when choosing a different sorting option or going back.